Merged updates from H3DUtil trunk, up to and including revision 127.
Added ReadWriteH3DTypes.h. A file containing a bunch of functions that can be used to write and read H3DTypes as binary data to and from a stream. Added in order to fix the DeviceLog of H3DAPI branch.

Image - Changed a variable type from byte to char

Matrix3d and Matrix3f - Optimized getScalePart function.

git-svn-id: https://www.h3d.org:8090/H3DUtil/branches/release1.0.1@128 875a7764-6f23-0410-b93c-dbbdadb1ea73