Added doxygen documentation (and actually an image) describing how DICOM files are loaded with and how their coordinate system relates to the Image voxel order.
LoadImageFunctions.cpp - Added checks to load DICOM data correctly on Windows when the locale is not English. This is done by checking the locale, and if not english test if "0,3" generates a value bigger than 0.1. In that case replace . with , in the string read from DICOM. Used only when checking ImagePatientPosition.

git-svn-id: https://www.h3d.org:8090/H3DUtil/trunk@212 875a7764-6f23-0410-b93c-dbbdadb1ea73